More portable shebang, from 7heo <7heo@mail.com>
Original commit message:
« The path /bin/sh for the shebang is more portable
than /usr/bin/env, according to Sven Mascheck:
http://in-ulm.de/~mascheck/various/shebang/#env
and:
http://in-ulm.de/~mascheck/various/shells/

Additionally, a space after the shebang first two
characters (#!) is said to be more portable:
http://in-ulm.de/~mascheck/various/shebang/#blankrequired

I fixed the test script shebang accordingly. »

It doesn't say the presence of a space makes it more portable; it says it was a rumor that could never be proven true, for some completely obsolete releases of UNIX. I did not apply that part.

Thanks!